TEST BRANCH for Symbola font subset
The existing webdrivercss tests were adapated to check the impact of
using a cut-down subset of the Symbola font.

To test:
- Install npm dependencies
- Create baseline screenshots `$ grunt test`
- Swap out the subset font:  `$ cp src/font/Symbola-subset.ttf build/font/Symbola.ttf`
- Run again `$ grunt test`

Check the folder test/style/screenshots/failed for mismatches. A small
number seem to be picked up due to a slight random horizontal shift -
these can be disregarded.

Edit test file test/style/specs.js to run either the full 30000 client
latex strings, or a sample of 500.