Add EXPERIMENTAL support for dynamic var symbol
The construction {{var:foo}} is now recognised and parsed as a separate
latex symbol - LrnDynamicVar.

Allows
- Initialisation of mathquill with dynamic vars in latex
- Addition of vars via .write() and .cmd()
- Copying of dynamic vars to clipboard
- Pasting of dynamic vars from clipboard
- Export of dynamic vars via latex()/text()

Copy/paste support means vars can be copied and pasted between CKEditor
source and mathquill if desired.

The var is treated as an immutable symbol - the cursor passes straight
over it and the var name cannot be edited.