[LRN] Refactor matrix to be based on Environments
Creates a separate symbol to handle the opening \begin{name} expression
that is used to create matrices and other environments - this means we
can remove the generic parser hack we previously used to detect the
start of a matrix.

This is in preparation for adding the array environment which is very
similar to matrix but requires a separate parser.

Also registers Environments to a separate object which is useful when
e.g. creating a new matrix via the cmd() public method.

LRN-20705